moved bootloader-config in krill-modules
